[fix] Change from converting data to integer to floating point

+++ src/hooks/useConnectionHooks.tsx
@@ -55,7 +55,7 @@ export default function UseConnectionHooksEffect() {
     };
 
     mqttClient.onMessageArrived = (message: Message) => {
-      const data = Number.parseInt(message.payloadString);
+      const data = Number.parseFloat(message.payloadString);
       if(isNaN(data)) return;
 
       switch (message.destinationName) {
